# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Leadership Engagement Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      parameters:
        - in: query
          name: upcoming
          schema:
            type: boolean
        - in: query
          name: type
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /events/{id}/leaders:
    get:
      operationId: listEventLeaders
      tags: [events]
      summary: List leaders attending an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of leaders at this event
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Person"
  /events/{eventId}/leaders/{personId}:
    post:
      operationId: addEventLeader
      tags: [events]
      summary: Add a leader to an event
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
        - in: path
          name: personId
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Leader added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventLeader"
        "409":
          description: Already added
    delete:
      operationId: removeEventLeader
      tags: [events]
      summary: Remove a leader from an event
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
        - in: path
          name: personId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "404":
          description: Not found

  # ─── Invitations ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Person:
      type: object
      required: [id, name, email, role, homeCity, homeState, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        title:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        role:
          type: string
          enum: [executive, secondary_leader, staff]
        homeCity:
          type: string
        homeState:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Event:
      type: object
      required: [id, name, location, city, state, startDate, eventType, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: string
        city:
          type: string
        state:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        eventType:
          type: string
          enum: [summit, conference, marketing, leadership, regional, other]
        leaderCount:
          type: integer
        inviteeCount:
          type: integer
        attendeeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    EventInput:
      type: object
      required: [name, location, city, state, startDate, eventType]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        location:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 1
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        eventType:
          type: string
          enum: [summit, conference, marketing, leadership, regional, other]
    EventUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        location:
          type: string
        city:
          type: string
        state:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        eventType:
          type: string
          enum: [summit, conference, marketing, leadership, regional, other]
    EventLeader:
      type: object
      required: [eventId, personId]
      properties:
        eventId:
          type: integer
        personId:
          type: integer
        addedAt:
          type: string
          format: date-time

    # ─── Invitation ──────────────────────────────────────────────────────────
...